+++ subsys/shell/modules/CMakeLists.txt
@@ -1,9 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
-zephyr_sources_ifdef(
-  CONFIG_KERNEL_SHELL
-  kernel_service.c
-  )
+add_subdirectory_ifdef(CONFIG_KERNEL_SHELL
+  kernel_service
+)
+
 zephyr_sources_ifdef(
   CONFIG_DEVICE_SHELL
   device_service.c
